+++ RELEASENOTES.md
@@ -92,6 +92,8 @@
         ([#9379](https://github.com/google/ExoPlayer/issues/9379)).
     *   Handle partial URIs in RTP-Info headers
         ([#9346](https://github.com/google/ExoPlayer/issues/9346)).
+    *   Fix RTSP Session header handling
+        ([#9416](https://github.com/google/ExoPlayer/issues/9416)).
 *   Extractors:
     *   ID3: Fix issue decoding ID3 tags containing UTF-16 encoded strings
         ([#9087](https://github.com/google/ExoPlayer/issues/9087)).